Move energy minimisation code
