+++ debug/optimisation_test_1.py
@@ -2,10 +2,9 @@ from multiprocessing import freeze_support
 
 import numpy as np
 
-from pyspinw.calculations.optimisation.energy_minimisation import ClassicalEnergyMinimisation, Free, Fixed
-from pyspinw.interface import couplings, axis_anisotropies
+from pyspinw.calculations.energy_minimisation import ClassicalEnergyMinimisation, Free, Fixed
+from pyspinw.interface import couplings
 from pyspinw.coupling import HeisenbergCoupling
-from pyspinw.gui.viewer import show_hamiltonian
 from pyspinw.hamiltonian import Hamiltonian
 from pyspinw.site import LatticeSite
 from pyspinw.structures import Structure
