+++ pyspinw/calculations/energy_minimisation.py
